Added a new enumerate, `VERIFY_STATUS', that returns three different values in according to the verification of the solution; `verify_solution()' now returns a `VERIFY_STATUS' and not a `bool'.
The functions `add_initial_conditions()' and
`Solver_Status solve_constant_coeff_order_1()' now are `const' (and not `static')
in the section private of the class `Recurrence'.

Added the methods `bool is_order_zero() const' and `void set_order_zero() const'.

Added the private methods in order to access to the data of
the class `Finite_Order_Info'.