Implemented the new methods `Expr::is_rational_scalar_representation()' and `Expr::is_rational_polynomial()' and removed the old `Expr::is_rational_polynomial()' that only tested the GiNaC's flag `GiNaC::info_flags::rational_polynomial'.
Moved the function `sqrfree()' from the file Expr.inlines.hh to Expr.cc:
now, before to call `GiNaC::sqrfree' on a expression, every occurence
of polynomial expressions for PURRS but not for GiNaC (e.g. log(2)) is
substituted with an arbitrary symbol and, after the decomposition with
`sqrfree', come back to the original expression.