In the classification's process rewritten the infinite order recurrence in normal form and used it in all successive computations.
Substituted some private data of the class `Infinite_Order_Info' with
the private data `associated_first_order_rec_'.
Added the methods in order to access to `associated_linear_rec_'
(publics in the class `Finite_Order_Info' and privates in the
class `Recurrence').

Several changes in all functions that used the methods that now are
no longer in the class `Infinite_Order_Info' because substituted by
`associated_first_order_rec_'.

Moreover, added the computation of a special case of infinite order
recurrence which has the upper limit of the sum equal to `n' and
the weight equal to `1'.